feat(trips) : Throw suitable errors and create separate file for enums
