+++ src/event/ride-accepted-event.ts
@@ -1,6 +1,6 @@
 export class RideAcceptedEvent {
   constructor(
     public readonly requestId: string,
-    public readonly acceptedAt: Date | null,
+    public readonly acceptedAt?: Date | null,
   ) {}
 }
