+++ src/types/trips.ts
@@ -1,0 +1,11 @@
+export enum TripStatus {
+  NOT_STARTED = 'not_started',
+  ON_THE_WAY = 'on_the_way',
+  REACHED_DESTINATION = 'reached_destination',
+  REACHED_PICKUP = 'reached_pickup',
+}
+
+export enum VehicleType {
+  TWO_WHEELER = 'two_wheeler',
+  FOUR_WHEELER = 'four_wheeler',
+}
